+++ pkg/runner/step_file.go
@@ -8,6 +8,7 @@ import (
 	"math/rand"
 	"os"
 	"path/filepath"
+	"strconv"
 	"strings"
 	"sync"
 
@@ -125,57 +126,86 @@ func (s *StepFile) Remove() error {
 }
 
 func (s *StepFile) ReadEnvironment() (*Environment, error) {
-	env, err := s.ReadDotEnv()
+	outputs, err := s.readAndConvertLines()
 
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("read env file: %w", err)
+	}
+
+	env := make(map[string]string)
+
+	for key, value := range outputs {
+		switch value.GetKind().(type) {
+		case *structpb.Value_BoolValue:
+			env[key] = strconv.FormatBool(value.GetBoolValue())
+		case *structpb.Value_NumberValue:
+			env[key] = strconv.FormatFloat(value.GetNumberValue(), 'f', -1, 64)
+		case *structpb.Value_StringValue:
+			env[key] = value.GetStringValue()
+		case *structpb.Value_NullValue:
+			env[key] = ""
+		default:
+			return nil, fmt.Errorf("read env file: key %q: cannot convert value type %q to string", key, structpbValueToTypeName(value))
+		}
 	}
 
 	return NewEnvironment(env), nil
 }
 
 func (s *StepFile) ReadValues(specOutputs map[string]*proto.Spec_Content_Output) (map[string]*structpb.Value, error) {
-	outputs, err := s.ReadKeyValueLines()
+	keyValues, err := s.readAndConvertLines()
 	if err != nil {
-		return nil, fmt.Errorf("reading outputs: %w", err)
+		return nil, fmt.Errorf("read output file: %w", err)
 	}
 
-	protoOutputs := map[string]*structpb.Value{}
-	for name, v := range outputs {
-		outputSpec, ok := specOutputs[name]
+	for key, value := range keyValues {
+		outputSpec, ok := specOutputs[key]
 		if !ok {
-			return nil, fmt.Errorf("output %q: unexpected output, remove from step outputs or define in step specification", name)
+			return nil, fmt.Errorf("read output file: key %q: unexpected output, remove from step outputs or define in step specification", key)
 		}
 
-		var outputJSON any
-		if err := json.Unmarshal([]byte(v), &outputJSON); err != nil {
-			return nil, fmt.Errorf("output %q: malformed, unmarshaling json: %w", name, err)
+		if err := s.checkOutputType(outputSpec.Type, value); err != nil {
+			return nil, fmt.Errorf("read output file: key %q: %w", key, err)
 		}
-
-		protoV, err := structpb.NewValue(outputJSON)
-		if err != nil {
-			return nil, fmt.Errorf("output %q: %w", name, err)
-		}
-
-		if err := s.checkOutputType(outputSpec.Type, protoV); err != nil {
-			return nil, fmt.Errorf("output %q: %w", name, err)
-		}
-
-		protoOutputs[name] = protoV
 	}
 
 	for name, o := range specOutputs {
-		if _, ok := protoOutputs[name]; ok {
+		if _, ok := keyValues[name]; ok {
 			continue
 		}
 		if o.Default != nil {
-			protoOutputs[name] = o.Default
+			keyValues[name] = o.Default
 			continue
 		}
-		return nil, fmt.Errorf("output %q: missing output, add to step outputs or remove from step specification", name)
+		return nil, fmt.Errorf("read output file: key %q: missing output, add to step outputs or remove from step specification", name)
 	}
 
-	return protoOutputs, nil
+	return keyValues, nil
+}
+
+func (s *StepFile) readAndConvertLines() (map[string]*structpb.Value, error) {
+	keyValues, err := s.ReadKeyValueLines()
+	if err != nil {
+		return nil, err
+	}
+
+	outputs := map[string]*structpb.Value{}
+
+	for key, jsonValue := range keyValues {
+		var outputJSON any
+		if err := json.Unmarshal([]byte(jsonValue), &outputJSON); err != nil {
+			return nil, fmt.Errorf("key %q: malformed, unmarshaling json: %w", key, err)
+		}
+
+		value, err := structpb.NewValue(outputJSON)
+		if err != nil {
+			return nil, fmt.Errorf("key %q: %w", key, err)
+		}
+
+		outputs[key] = value
+	}
+
+	return outputs, nil
 }
 
 func (s *StepFile) checkOutputType(want proto.ValueType, have *structpb.Value) error {
@@ -201,6 +231,8 @@ func structpbValueToTypeName(value *structpb.Value) string {
 		return "string"
 	case *structpb.Value_StructValue:
 		return "struct"
+	case *structpb.Value_NullValue:
+		return "null"
 	default:
 		return "unknown"
 	}